Fixed comparison of empty repeated/map extensions.
Repeated/map extensions are semantically equivalent to an extension that is not present at all. We had code paths that were treating them differently, which led to incorrect results.

In particular, we were considering `{.repeated_ext = []}` to be different from `{}` when comparing with `upb_Message_IsEqual()`.  This change fixes this bug so that they will be considered equivalent.

PiperOrigin-RevId: 700787434