Rust: remove references to `crate::` from generated code
This is necessary for the generated code to work correctly even when it's
placed in a module instead of directly in the crate root. Now we can finally
delete the last of the gencode post-processing from the protobuf_gencode crate.

To get this to work properly I had to update the `Context` object to keep track
of the current module depth. This way we know how many `super::` prefixes we
need to prepend to an identifier to get back to the top level.

I had to some refactoring of our naming helper functions to get everything
working properly:
 - Deleted `GetCrateRelativeQualifiedPath()`, since it seems simpler if we just
   always provide unambiguous paths. I added some new `RsTypePath()` overloads
   as a replacement.
 - Made `RustModuleForContainingType()` a private implementation detail of
   naming.cc, since the `RustModule()` functions are more user-friendly and
   accomplish the same thing.
 - Moved the logic for prepending super:: or the foreign crate name into
   `RustModuleForContainingType()`. This way, all the helpers that call that
   function automatically pick up the behavior we want.

PiperOrigin-RevId: 703555727