Rust: set `ignore` on code example in comment
I found that `cargo test` fails without this, because even with `no_run`, Cargo
will still try to compile the code. It would be nice to have a working example
here, but unfortunately it's non-trivial to set that up since we would need a
real generated proto.

I also tweaked cargo_test.sh to run the doc tests so that we have CI coverage
for those.

PiperOrigin-RevId: 703611549