Raise ParseError for non-numeric strings in numeric fields in Ruby and PHP JSON parsing.
This fixes Ruby and PHP to be conformant with the Protobuf's JSON spec. Note this fix is not accompanied by a major version bump for Ruby or PHP, but was pre-announced in https://engdoc.corp.google.com/eng/doc/devguide/proto/news/2024-11-07.md#ruby-and-php-errors-in-json-parsing and landed as a warning in 29.x.

PiperOrigin-RevId: 704855202