hpb: support hpb::RepeatedField<T> inside extensions (GetExtension)
Before this change, hpb had no way of returning repeated fields (that are extensions) -- they were incorrectly treated as pure scalars (int32 vs repeated<int32>).

We rectify this hole and now return RepeatedField<T> for a given T.

This CL also cleans up the `if constexpr` special casing we were performing inside GetExtension and delegates that to the UpbExtensionTrait.

PiperOrigin-RevId: 706789273