hpb: Introduce UPB_EXT_PRIMITIVE to flesh out all remaining scalars for extensions
In this CL, we add the macro UPB_EXT_PRIMITIVE.
The template specializations are practically identical sans the CppType and UpbFunc called, so we now consolidate via this macro.

Added support for uint32/64, float/double, and bool.
Getting and setting exts of ^ in hpb should all work, and fetch the proper default value as well (if provided in the .proto).

PiperOrigin-RevId: 707897721