Add python_api_test to ensure that GitHub runners runs with expected Python backend.
In v21.x, we announced our intent to flip the default implementation of
protobuf python from pure python to upb.
- https://github.com/protocolbuffers/protobuf/releases/tag/v21.0

We also documented in docs and readme that
PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION env var can be set to override the
python implementation used.
- https://protobuf.dev/reference/python/python-generated/#sharing-messages

Our python binaries are released with UPB enabled:
- https://github.com/protocolbuffers/protobuf/blob/main/python/dist/BUILD.bazel#L380

By in Bazel tests, we only use upb if it can successfully be imported:
- https://github.com/protocolbuffers/protobuf/blob/1c29f34b24eafe4a362ce075c12c62c614fad199/python/google/protobuf/internal/api_implementation.py#L54
- https://github.com/protocolbuffers/protobuf/blob/1c29f34b24eafe4a362ce075c12c62c614fad199/python/google/protobuf/internal/api_implementation.py#L100

Otherwise, we fall back to pure python, even if the env var is set to upb.
Since we don't currently link in upb for Bazel, this means the default is
still pure python in practice and the environment variable is not respected.

This change does not enable UPB, but adds a python_api_test to ensure that the
Bazel runners is running with the python backend that we expect.

#test-continuous

PiperOrigin-RevId: 711436307