Don't clear hasbits for optional fields which have `mutable_` accessors.
Though uncommon, it is valid to obtain a mutable reference to a proto field, clear that field, then modify the cleared field via the reference. If this field has hasbits, clearing the field will clear the hasbits, but modifying the field through the reference to the underlying type will not. This can lead to a present field missing its hasbit.

PiperOrigin-RevId: 720363816