move attribute after enum class to fix clang warning
Building with clang produce the following error:
text_format.h:55:1: warning: attribute 'visibility' is ignored,
place it after "enum class" to apply attribute to type declaration [-Wignored-attributes]

The warning occurs because attributes like
`__attribute__((visibility("default")))` are ignored when placed before
an `enum class`. This is due to how attributes are parsed in C++.

Moving the attribute after `enum class` ensures it is applied
correctly to the type declaration.

Signed-off-by: Clément Péron <peron.clem@gmail.com>