Use `toUnsignedString()` to convert `uint32` and `uint64` to strings.
This method has existed since Java 8, released in 2014.

Also adjust the tests so that they verify unsigned integer values that are too big for the signed counterpart (for example, 4294967295 which fits in `uint32` but not `int32`).

We don't use the parallel `Long.parseUnsignedLong` because the existing code handles inputs like `"1.0"`.

PiperOrigin-RevId: 723945649