Allocate enough memory for the InputOver2G test.
This test was previously causing a pointer to be taken 2GB out of bounds,
which is UB and can cause problems for tools. Fix the problem by using
a suitably sized buffer instead. With most allocators this will just
result in pages being mmap()ed without being faulted so this test won't
actually end up using 2GB of memory.

PiperOrigin-RevId: 724172825