Tolerate message set extensions encoded as normal deliminited submessages
Currently upb will only tolerate parsing a message set extension if it's encoded
in the bespoke message set format. If it's encoded as a normal submessage it'll
be relegated to the unknown field set.

As not all encoders are message set aware, tolerating both forms when decoding
will improve compatability.

PiperOrigin-RevId: 724363865