Add ifdef guards to prevent tests from consistently failing ASAN and UBSAN.
In the GenCodeMapMissingKeyDeathTest we deliberately trigger an invalid map
access to try to understand its behaviour. This test case was added in https://github.com/protocolbuffers/protobuf/commit/27ee74e6c6f55759eab3dcb7f8bbdce9891bd433.

This undefined behaviour was gated by an ABSL_DCHECK and will consistently
crash in debug mode. In optimized mode, this is invoking undefined behaviour
and having a test case for it is meaningless. (We have seen it crash or run
without errors depending on sanitizer and other compiler flags).

I've removed this test case altogether in NDEBUG mode.

PiperOrigin-RevId: 733919241