Document that `RepeatedPtrField` guarantees pointer stability
`RepeatedPtrField` stores an array of pointers and makes no effort to hide that
fact, but as of today it does not explicitly guarantee that elements have
stable addresses. This guarantee is reasonable and is more-or-less implied by
the name already, so let's make it explicit.

Fixes https://github.com/protocolbuffers/protocolbuffers.github.io/issues/241.

PiperOrigin-RevId: 739246630