Eliminate an signed/unsigned arithmetic footgun in enum range validation.
In C++ the result of an (int + unsigned int) is implicitly promoted to an
unsigned int. This is surprising! I doubt that was the intention of the
original enum validation code.

Because we use a (start, length) tuple to validate enum range, in parse loop we
compare that start <= val < start + length. However start is int16_t and range
is uint16_t. The result of (start + length) is an unsigned value.

This means that if we have a closed enum range of {-5, -4, -3}, we will have:
- start  = -5
- length = 3

int32_t{-5} + uint32_t{3} = 4294967294, as seen here:
https://cpp.sh/?source=%2F%2F+Example+program%0D%0A%23include+%3Ciostream%3E%0D%0A%23include+%3Cstring%3E%0D%0A%0D%0Aint+main()%0D%0A%7B%0D%0A++std%3A%3Astring+name%3B%0D%0A++std%3A%3Acout+%3C%3C+(int32_t%7B-5%7D+%2B+uint32_t%7B3%7D)+%3C%3C+std%3A%3Aendl%3B%0D%0A%7D

We are in trouble! But as it turns out we got lucky, because we immediately do
a comparison with it in `val < start + length`. Because we are doing a
comparison between signed and unsigned types, val gets promoted to an unsigned
type, and due to two's complement, it just so works out that
uint32_t{-1} > uint32_t{-2}.

All of this is working today but it's hanging by a thin thread and relies on a
lot of assumptions about field width. Signed integer types using two's
complement isn't even ratified in the standard until C++20:
https://stackoverflow.com/questions/57363324/ramifications-of-c20-requiring-twos-complement

Long story short I think it'll be less headache inducing if we just spell out
the types explicitly here and keep everything signed. start and length are each
16 bits anyway and we shouldn't have any trouble casting them to signed int32s.

PiperOrigin-RevId: 740444524