Replace `ABSL_ATTRIBUTE_UNUSED`  with `[[maybe_unused]]`
Currently `ABSL_ATTRIBUTE_UNUSED` is not set for a least
MSVC 19.29.30158.0, leading to warnings or error in some cases.

`[[maybe_unused]]` is now part of the standard and widely adopted.
Using this attribute makes the compilation portable,
resolving issues for this version of MSVC.

See the deprecation notice of `ABSL_ATTRIBUTE_UNUSED` in
`abseil/abseil-cpp`:
https://github.com/abseil/abseil-cpp/blame/master/absl/base/attributes.h#L552-L566

Signed-off-by: Julien Jerphanion <git@jjerphan.xyz>