Make Rust crates add a `-release` suffix for all non-RC versions
The idea behind this is that Cargo's lexicographic ordering will understand
that `4.31.0-release` is newer than all release candidate versions such as
`4.31.0-rc.1`. Hopefully it's also intuitive for humans that the `-release`
version is the "real" release, but we need to keep some kind of hyphen suffix
on the end to indicate that we are not committing to a stable API yet.

PiperOrigin-RevId: 740928112