Keep unimportant fields when possible when compressing the fast-parse table.
Previously, if an unimportant field was in the first half of the fast-parse table during compression, and there was no field in the merging slot from the second half, the field would be discarded. This is a small optimization which will allow keeping unimportant fields whenever possible while compressing the fast-parse table.

PiperOrigin-RevId: 740933470