Cache-friendly `RepeatedPtrField`: Plumb allocation traits to `InternalExtend`
The element size is required to allocate contiguous buffers for future elements.

The element alignment is optional in the current de facto state of the code base, because `Arena::CreateArray()`, which is used to create buffers, happens to always align to 8 bytes, which is sufficient for all currently possible element types. However, the alignment is plumbed anyway here, and subsequent CLs assert (in debug builds only) that created buffers are properly aligned.

PiperOrigin-RevId: 741308845