Fetch benchmark and googleapis dependencies through bzlmod
This required downgrading our use of the ads API, which we use only for
benchmarking. The most recent googleads version on BCR only supports v18, and
we can't easily use this one because it requires Bazel 7.2.1 or higher. So
instead we need to use an earlier version that only includes v17. I think this
is fine, though, because the v17 API is still supported and it should be easy
to upgrade once we have upgraded our Bazel version.

PiperOrigin-RevId: 741659742