hpb: cleanup DeductionBlocker -> DeductionBarrier
DeductionBlocker can sound like this blocks deduction. The intent is to have these varadic args greedily match and *force* deduction.

Therefore, DeductionBarrier is a better name, as this blocks every template param on the right side of the variadic args from being manually specified.

Note that int&... was also specially chosen, as it's very unlikely to accidentally specify a reference to an int with static storage.

PiperOrigin-RevId: 745199015