Add a bunch of optional changes to csharp code generation
- strip out features we don't need
- simplify the generated code
- add in Unity serialize field attributes
- improve the hash functions
- support enabling nullable fields

These changes are all off by default, and many/most of them break the unit tests and such if you enable them, because of course, the unit tests test all the features and we’re deleting them. With all of these things enabled, though, we save about 20% on the generated code size in our project.