Fix link errors when building `libprotoc` using latest llvm-mingw toolchain
Looks like Clang in MinGW mode doesn't add explicit
instantiations of a template function (defined in a separate
translation unit) to the overall export set of the library,
despite the enclosing class having `PROTOBUF_EXPORT` attribute.

This happens when building `libprotoc` with LLVM-MinGW toolchain,
e.g., while trying to build/install protobuf via vcpkg for one
of the `*-mingw-dynamic` triples.

So, add another `PROTOBUF_EXPORT` to the explicit instantiation
declaration for `RepeatedPtrFieldBase::MergeFrom<std::string>`.
This won't hurt in either case, but resolves the link issue for
llvm-mingw builds of the `libprotoc` library.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>