Move LazilyBuildDependenciesTest into its own .cc file and CMake test binary
One of the test cases involves observing the lazy loading behavior of the
generated descriptor pool. Since this pool is a global, it can be influenced by
other unrelated tests, and whether the test succeeds or fails depends on the
test ordering. Moving the test into its own binary ensures that it will pass
regardless of test ordering.

Fixes #21431.

PiperOrigin-RevId: 752313710