Fixed Python source wheel breakage caused by https://github.com/protocolbuffers/protobuf/pull/21423
The Python source wheel is pulling in `decode_fast.c`, leading to compilation errors when it is not enabled.

While we could try to prevent this file from being in the source wheel at all, that is probably a short-term solution.  If fasttable support is ever ready for prime time, we will need to be able to decide at build time whether to enable it or not, because it's not supported on all platforms.

#test-continuous

PiperOrigin-RevId: 752479765