hpb: Add plumbing for detecting cpp backend in protoc-gen-hpb.cc
Just a stub, so we'll abort if we detect c++ right now.

Note the change from `const Options&` to `Options` for `Context`. Since `const T&` binds to temporaries, we were creating an Option{}, and immediately throwing it away, causing a use-after-free-from-stack. We'll opt to copy by value to rectify the above.

PiperOrigin-RevId: 753183442