Make ExtensionRegistry reject field numbers that are out of range.
Usually extension numbers come from descriptors that have previously been validated by `protoc`, so we don't expect this to affect real systems in the wild, but this edge case was noticed by fuzz tests.

This is not a full validation of the field number -- for example, we do not verify that the extension number is from one of the extension ranges declared in the `.proto` file.  (This would not be possible to verify currently, as MiniTables do not know the extension ranges.)  We verify just enough to remove the risk of accepting invalid wire format.

PiperOrigin-RevId: 755989392