Change ChangeOneof into void return type and let it initialize oneof fields.
We can remove a runtime branch for the far more common "singular field" case if
we hoist the oneof field initialization logic into the ChangeOneof helper
function. ChangeOneof is called whenever we've hit the branch that indicates
the message is a oneof field, but if we put the initialization logic outside of
it, we essentially need to do the check again.

PiperOrigin-RevId: 757602106