Use `DescriptorPool` to cache whether Java needs an outer class suffix.
Memoizing this information in the descriptor pool will help us unwind a lot of complexity.  Right now most of our Java naming APIs are burdened by the fact that we cannot compute the file class name without performing an O(n) search of all classes, enums, and services within the file, leading to O(n^2) overall behavior unless we can cache this search result.  But once we need a cache, we can no longer use simple functions like `ClassName(const Descriptor*)`, it instead has to be `cache.ClassName(const Descriptor*)`, requiring the cache to be passed around between components.

The need for a cache has caused great complexity in our APIs.  The APIs in http://google3/third_party/protobuf/compiler/java/names.h are simple but have `O(n)` behavior.  As a result, the code generator does not actually use these APIs, but instead uses `ClassNameResolver`, which caches the results.

Memoizing this info in the DescriptorPool will let us unwind a lot of this complexity, and unify the APIs.

PiperOrigin-RevId: 762128834