Disabled all fast parsing functions and enabled fuzzing of fasttable.
From this point forward we will expect all fasttable functionality to be
fully correct and fuzzed. This formally marks the transition from
the prototype stage to developing the prod version.

However we will still hide this functionality behind a build flag for
now, because including fasttable dispatch creates a slight performance
regression until we are actually doing some fast parsing.

PiperOrigin-RevId: 763248497