Use blanket implementations for `*GetRawMessage*` traits
This CL sets things up so that `CppGetRawMessage` and `UpbGetRawMessage` are
implemented in the gencode only on view proxies, and the corresponding `Mut`
traits are only implemented in the gencode on the mut proxies. The appropriate
remaining implementations on owned messages and mut proxies are then handled by
blanket implementations in the runtime. This cuts down on the size of the gencode a
little bit.

PiperOrigin-RevId: 764833547