Propagate .pyi files via direct_pyi_files and transitive_pyi_files
Per reviewer feedback, we here propagate the paths to the `.pyi`
files via the `direct_pyi_files` and `transitive_pyi_files` members
of `PyInfo` which are available starting in `rules_python` 1.1.0.

Since this means that the files are not automatically in the runfiles
of the `py_proto_library`, tooling to run mypy will need to be
adjusted to add what is needed from these provider fields to the
runfiles if that is what is wanted.