Fasttable: Added support for repeated fixed fields.
We use a design where a single function handles both repeated and non-repeated, but calls cardinality helpers that reduce to nothing for the non-repeated case.  This lets us share much of the repeated field handling code between different field types, and reduces the number of fast parsing functions we need to write.

For the moment we avoid calling any non-inline function to avoid spilling registers in uncommon paths like "arena is out of memory."  We either handle the case entirely inline or we return to fall back to the MiniTable parser.

We can play around with some of these implementation choices once the parser is complete and correct.

PiperOrigin-RevId: 767323516