Rust: implement ProxiedInRepeated with a blanket implementation for upb
Most of the method implementations ended up being identical to the gencode
implementations. The only ones that had to be significantly different were
`repeated_get_unchecked` and `repeated_get_mut_unchecked`. I had to make them
rely on the `UpbTypeConversions` trait instead of using methods on the
generated view and mut proxies.

The motivation for this is to try to implement as much as possible in the
runtime instead of in generated code. That should help mitigate the downsides
of needing to accommodate stale gencode, because it will allow us to add
features or modify behavior in a way that takes effect even if the gencode is
not updated.

PiperOrigin-RevId: 767596134