Remove forward declarations from compiler/java/context.h
When we have the combination of C++ 20, Clang <= 14, and a recent Abseil
release, the compiler errors out on a `absl::flat_hash_map` with an incomplete
value type. One place where we hit this is in a couple of maps in
compiler/java/context.h where the value type is a forward declaration.

This change works around that problem by removing the forward declarations and
instead including the header that defines them. This is a bit more complicated
than it sounds, though, because it requires breaking up some build targets to
avoid dependency cyles.

PiperOrigin-RevId: 773840460