Generate .pyi files in py_proto_library (#10366) (#21567)
We here adjust py_proto_library to produce `.pyi` files along with the `.py` files it already generates.  This achieves the same effect as the gRPC py_proto_library.

Per reviewer feedback, we here propagate the paths to the `.pyi`
files via the `direct_pyi_files` and `transitive_pyi_files` members
of `PyInfo` which are available starting in `rules_python` 1.1.0.

Since this means that the files are not automatically in the runfiles
of the `py_proto_library`, tooling to run mypy will need to be
adjusted to add what is needed from these provider fields to the
runfiles if that is what is wanted.

Closes #10366.

Closes #21567

COPYBARA_INTEGRATE_REVIEW=https://github.com/protocolbuffers/protobuf/pull/21567 from dws:py_proto_library-pyi 37fbdc086c3b636517432230b0ba97933ac02fc6
PiperOrigin-RevId: 778116905