Upgrade Clang on Windows test runs and work around MSVC error (#22467)
Most of our Windows test runs are currently failing with an error message about
the Clang version being too old. This seems to be an occasional
[issue](https://github.com/actions/runner-images/issues/10001) with the GitHub
Windows runners, and the suggested workaround is to use `choco` to upgrade LLVM.
We should revert this change as soon as the GitHub runners are fixed.

This change also includes a tweak to descriptor.cc to work around a mysterious
"internal compiler error" with MSVC.

PiperOrigin-RevId: 773825728