hpb: make ExtensionRegistry backend-aware
When the backend is upb, the registry_ var will be initted and supplied an arena. One can also add an extension, and get the generated registry (under the hood, uses upb linker arrays). An empty registry is also possible here.

The cpp backend doesn't really have a notion of an ExtensionRegistry - it always uses the generated one. Since the C++ backend can't construct an ExtensionRegistry directly, hpb functions must be called with hpb::ExtensionRegistry::generated_registry().

hpb::ExtensionRegistry::EmptyRegistry won't even exist if users are on the C++ backend.
PiperOrigin-RevId: 786261025