Expand `embed_edition_defaults` encoding to byte types.
While there is support for base64, it requires a runtime cost to decode
back into the bytes, so adding support to write out either hex or decimal
byte values to the encoding. This allows the values to be directly used for
datatypes that are more "byte buffer" than a "string".

Also make the rule's implementation return the output file via the default file
provider.

PiperOrigin-RevId: 786338818