Rust: remove `DEP_UPB_VERSION` check in codegen crate (#22764)
This check was meant to enforce that the version of the protobuf crate exactly
matches the version of the codegen crate, but this is not strictly necessary.
Tonic would also like to re-export our crate from the `tonic-protobuf` crate,
and the `DEP_UPB_VERSION` check is causing problems with that since the
environment variable is set only for crates that directly depend on the
`protobuf` crate (see
[discussion](https://github.com/hyperium/tonic/pull/2320#discussion_r2228822853)).

While I was at it I also removed the code in the protobuf build script that
sets the `DEP_UPB_INCLUDE` variable, which is unnecessary now that we no longer
generate any C code which would need to know where to find upb's headers.

PiperOrigin-RevId: 786760264