Always emit hasbit index for fields regardless of whether hasbits are used in the message.
In the future, we would like to conditionally assign hasbits to fields based on PDProto profiles/etc. This means we won't be able to tell by the type of a field alone whether a field uses hasbits.

The TCParser currently assumes hasbits for field types which are eligible for hasbits. We want to emit -1 here for these fields which don't use hasbits but are eligible based on their type so the TCParser can correctly ignore them.

As written currently, if a message does not use hasbits and contains a field which is eligible for hasbits but does not have them assigned, that field entry's hasbit index will be 0, which the TCParser will not be able to distinguish from a field which was assigned hasbit index 0.

PiperOrigin-RevId: 788586969