Upgrade rules_rust to 0.63.0 so that we can build with Bazel 9 prereleases
I tried building `//rust/...` with the most recent Bazel 9 prerelease and found
that it succeeds once we upgrade rules_rust. This is not critically important,
but it seems nice to be able to try out Bazel 9 in advance.

Unfortunately we can't upgrade rules_rust in WORKSPACE without breaking
workspace builds, but I think this is probably fine since that is already
deprecated anyway. It's also broken on Bazel 8, so presumably we will drop
support for Rust WORKSPACE builds entirely once we drop Bazel 7.

PiperOrigin-RevId: 792307568