Use Rust 1.79 for Bazel and nightly for Cargo
We need to use a recent nightly release to be able to use the `--workspace`
flag to `cargo publish`. Our minimum supported Rust version is 1.79,
so I think we should test that version with Bazel to make sure we stay
compatible. Our Bazel build is currently using 1.86 since that is the default
from rules_rust, but this change explicitly selects 1.79.

I also tried having Bazel use a more recent version, but unfortunately
rules_rust is broken for 1.87 and up, due to name mangling changes in rustc:
https://github.com/bazelbuild/rules_rust/issues/3459

As soon as Rust 1.90 is released, we can start using that with Cargo instead of
nightly.

PiperOrigin-RevId: 793660982