Add new protobuf_tests crate
We currently have a number of tests that we exercise with Bazel only, so the
goal of this change is to remedy that by making sure we can run them
successfully with Cargo as well.

Most of these tests operate on code generated by protoc, and we want anyone to
be able to run the protobuf crate's tests without having protoc installed. I
therefore created a new crate called `protobuf_tests` which is solely for
testing and will not be published to crates.io.

The plan is for almost every test under `rust/test/` to eventually go in this
crate, but for now I just started with `proto_macro_test.rs`.

PiperOrigin-RevId: 794286695