Fixed mypy errors by setting `__slots__` to empty in `.pyi` files.
The `__slots__` are redundant, since they duplicate the members we already have. Also, they cause problems with mypy: https://github.com/protocolbuffers/protobuf/issues/21691

We have to leave an empty `__slots__ = ()` in place, as opposed to removing it entirely, because that's the only way to ensure that pytype will continue to error out if an unknown member is read or written.

PiperOrigin-RevId: 795107868