Move implementation of UpbTypeConversions from gencode into the runtime
The implementation of this trait is tied somewhat closely to the representation
of the C `upb_MessageValue` union, so it seems useful to have it in the runtime
where we can make sure it stays up to date with any changes to upb.

Messages and enums need different blanket implementations, and Rust does not
allow more than one blanket implementation of the same trait. I worked around
this by adding a type parameter on the `UpbTypeConversions` trait. The
parameter does not do anything except act as a tag, so I named the tags
`MessageTag`, `EnumTag`, and `PrimitiveTag`.

The one place where I ran into trouble is that there was no generic way to go
from a `RawMessage` to a mut proxy. I worked around this by having the
generated code for mut proxies implement `From<MessageMutInner<'_, Message>>`.
I realized that this was actually useful for the C++ kernel as well, since it
made it possible to move the implementation of `MessageMutInterop` into the C++
runtime.

PiperOrigin-RevId: 799683133