Move toward a clean separation between Rust generated code and the unstable upb API
This CL updates upb.rs to explicitly list out everything from upb that we
currently need to expose to the gencode. For each item on this list, we need to
eventually fix the gencode so that it does not directly reference it, to allow
upb to evolve without breaking old gencode.

I added a `debug_string()` helper function and updated the gencode to call that
instead of `upb::debug_string()`.

I also deleted some deprecated methods on `UpbGetMessagePtr` and
`UpbGetMessagePtrMut` just because that was easy to do.

PiperOrigin-RevId: 799730937