Move the implementation of the `Serialize` trait into the runtime
The main motivation for this is to make it easier to accommodate stale upb
gencode, but I did this for both the upb and C++ kernels since it was
straightforward to do both.

While I was at it, I also updated `upb::wire::encode` to take a `MessagePtr<T>`
instead of a raw message and a MiniTable pointer. This removes the possibility
of a mismatch between the message and MiniTable, allowing the function to be
marked safe.

PiperOrigin-RevId: 800041977