Let IsFieldMicroString not crash or have otherwise undefined behavior.
IsMicroString will crash in debug mode when the provided field descriptor
is not a string. There seems to be nothing in the contract to prevent
the user from passing a non-string FieldDescriptor.

IsInlined seems to give a precedent of what we'd do in this situation.
Namely, instead of ABSL_DCHECK, we should just return false and not
crash (or invoke UB).

PiperOrigin-RevId: 800293756