Move as much as possible of mini table generation into the runtime
For enums, this CL changes things so that their generated code implements a
trait that just exposes the mini descriptor. From there, the runtime does the
work of actually building the mini table in a blanket implementation of the
`AssociatedMiniTableEnum` trait.

I thought about doing something analogous for messages, but the fact that we
need to accommodate cyclic message graphs made this seem too complicated to be
practical. For messages I instead just added `build_mini_table()` and
`link_mini_table()` functions so that the generated code can call those instead
of the lower-level upb functions.

As a side cleanup, I removed the implementations of `AssociatedMiniTable` on
view and mut proxies, because I realized that these were not necessary. Any
code that needs the mini table can easily get it from
`Self::Proxied::mini_table()`. This is not super important but does cut down on
the generated code size a little bit.

PiperOrigin-RevId: 802324461