Move the implementation of `ProxiedInRepeated` fully into the upb runtime
Previously `ProxiedInRepeated` was implemented with a blanket implementation
only for messages, but this CL extends that blanket implementation to work for
all repeated field types.

The one thing that was tricky about this is that the implementations of
`ProxiedInRepeated::repeated_copy_from()` are different enough that I couldn't
see a clear way of unifying them into a single implementation. Instead, I added
a `copy_repeated` function on the `UpbTypeConversions` trait and moved the
logic there, so that messages, enums, numeric types, and string types can all
have different logic there.

PiperOrigin-RevId: 802568222