Stop forcing linker array entries to be retained.  They will now be GC'd like normal, but will be iterable if retained.
There was no need to force them to be retained.  Extensions only need to be retained if they are referenced, and normal linker GC can tell us if they are referenced.

This will make the ExtensionRegistry linker array agnostic to how extensions are packaged into TUs.  However, TUs are still relevant if we want to perform "implicit weak" static tree shaking, because any extensions that are colocated in the same file will cause all referenced sub-message types to be pulled into the link.  But this is no worse than what we get for regular messages -- for implicit weak we should always put every message and every extension into its own TU.

PiperOrigin-RevId: 803130581