Move the `Parse` trait out of gencode and into the runtime
As part of this, I also removed the inherent methods `parse` and
`parse_dont_enforce_required` from generated messages. This changes the API in
a way that requires fixes for code that's not using our prelude (requires
adding `use protobuf::Parse`).

The blanket implementation of the trait is the same for both the C++ and upb
kernels, so I just put it directly in codegen_traits.rs.

PiperOrigin-RevId: 804967515