Add type tags for view and mut proxies
We're not using these for anything at the moment, but I think it's a good idea
to put them in the gencode now just in case we ever need them.

One thing I have been thinking about in particular is that at some point in the
future we may need generic code to behave differently depending on the vintage
of gencode it is working with. For example, we might want to enable a new
feature on gencode that is new enough to support it, while it leaving it safely
disabled on older gencode. If we ever need to do that, we can create new tags
such as `MessageV2Tag` and thereby distinguish between `EntityType<Tag =
MessageV2Tag>` and `EntityType<Tag = MessageTag>`.

If we ever want to move a trait implementation from the gencode to the runtime,
or vice versa, I think this will also help provide a way to ensure that the
trait is implemented exactly once per type.

Without this CL, we can already do this for owned messages and enums, but this
CL just enables the same possibility for view and mut proxies.

PiperOrigin-RevId: 805412181