Move as much as possible of mini table generation into the runtime
I tried to move `AssociatedMiniTable` and `AssociatedMiniTableEnum` completely
out of the gencode and into the runtime, but this turned out to be impractical
because a trait blanket implementation cannot create a per-monomorphization
static variable. Messages are also especially complicated because of the need
to handle cycles when linking. So instead, I kept the trait implementations in
the gencode but set them up to call helper functions such as
`build_mini_table()` that are at a higher level than the raw upb functions.

I did remove the implementations of `AssociatedMiniTable` on view and mut
proxies, because I realized that these were not necessary. Any code that needs
the mini table can easily get it from `Self::Proxied::mini_table()`. This is
not super important but does cut down on the generated code size a little bit.

PiperOrigin-RevId: 805592273