Loosen version check to allow for checked-in gencode
The necessary logic was a little bit tricky to implement in a const context but
I seem to have gotten something working that does what we want.

The logic now considers gencode to be compatible as long as it's not newer than
the runtime, and also pre-4.33 gencode is always considered incompatible since
4.33 will (hopefully) be our first stable release.

One thing to note is that this logic technically allows compatibility across
major versions, which we may or may not be what we want. But if/when we ever
increment the major version, we can update this logic accordingly.

PiperOrigin-RevId: 805905247