Fixed a fuzz bug by properly rejecting MiniTables with improper map fields at schema creation time.
Specifically, any enum used as a map value must contain a zero value.  Before we were not enforcing this, which could lead to the wrong enum value being serialized if we parsed a map entry with no value.

This change should have no real effect in prod, because protoc will already reject enum fields that break this invariant.

PiperOrigin-RevId: 806683268