Pass arena down to `RepeatedPtrField` accessors in generated code.
Future optimizations to `RepeatedPtrField` will make `GetArena()` more expensive. Reading the arena pointer directly from the message will be cheaper, so we should pass arena pointers down to `RepeatedPtrField`s from the containing message whenever possible.

We define `Internal(Add|MergeFrom)WithArena` methods which take an arena pointer as an argument, which is required to be the arena pointer from the parent message. These specializations are for internal use only, and function the same as their public counterparts, aside from avoiding a call to `RepeatedPtrField::GetArena()`.

PiperOrigin-RevId: 810533181