Refactor `MapFieldBase::ReflectionPayload` to be a class.
This abstracts away the representation of members of this class and provides access to them via accessors/mutators. In a future cl, we will have to wrap the `RepeatedPtrField` member of this class, and having an accessor to the field rather than exposing it directly contains the blast radius of that change to just this class.

PiperOrigin-RevId: 810699795