Add bzlmod support for system_python.
#test-continuous

This creates a module extension for system python so that it can be used in bzlmod builds.  The pip_parse setup (dynamically building a new repo) doesn't work in the new bzlmod model, so we use the regular pip extension pointed at system_python's interpreter.

Remaining work to get off WORKSPACE:
* The pip extension quickly fails in the case where there's no python available, unlike our system_python setup
* Our windows upb tests were never actually running python, and system_python doesn't work (and breaks under bzlmod due to the above issue)
* Some of the upb wheel tests still don't work on bzlmod yet because they depend on some very custom http_archive dependencies.

Closes #18750
Closes #23313
Closes #23307

PiperOrigin-RevId: 810995270