Refactor `FieldArenaRep` to have a default implementation and add a `Get` method to recover the wrapped type, and unify object construction code under `CreateArenaCompatible`.
In the future, some field types will be wrapped in a containing type when they are individually allocated. To simplify code which needs to access the wrapped type, we will define a `Get` method on the `FieldArenaRep` helper struct, which will be defined by each field type overload to recover the wrapped type.

PiperOrigin-RevId: 811954764