Add gencode smoke tests (#23547)
This creates a new //compatability/smoke directory which has a simple unit test and checked-in gencode which was produced by a given protoc version.

Included is a script that you can run with a protoc release number, which will fetch the released protoc linux-x64 binary, run it on the test proto, add the code to a subdirectory and hook up the unit test to that generated code.

If this setup looks good we can quickly add a number of additional versions; I have already validated it works with a wider range but this PR only includes one version to show the setup and focus the initial review on the core things that aren't about the (naming, directory structure, etc)

Closes #23547

COPYBARA_INTEGRATE_REVIEW=https://github.com/protocolbuffers/protobuf/pull/23547 from protocolbuffers:esrauch-stalegencodetest 7b89e900cc375614214afcc7b8efe25cc47a472c
PiperOrigin-RevId: 812933488