Allocate `FieldArenaRep` type in `Arena::DefaultConstruct` when using arenas.
This change is prerequisite to supporting split fields with removed arena pointers. Split field types which have removed arena pointers have a larger representation on the arena which holds a copy of `InternalMetadata` from the parent message, since they need to be able to find a nearby `InternalMetadata` within a 32-bit offset. This is exactly how directly arena-allocated field types with arena offsets will work, but this logic is currently only implemented in `DoCreateMessage`. We will do the same trick here to construct this larger type when allocating in `Arena::DefaultConstruct`, which is the API split fields use.

PiperOrigin-RevId: 814896561