Add `TaggedInternalMetadataResolver` type which can store data in the unused bits of an offset.
Offsets will be a multiple of the minimum of the alignment of the containing class and `InternalMetadata`, typically 8 bytes on 64-bit platforms. This gives us a few bits at the beginning of the offset to use for other purposes, which will be particularly useful in the future for `RepeatedField` to hold the SOO state.

This cl provides an interface for packing tag bits in `InternalMetadataResolver`s without paying a runtime cost when tags aren't used.

PiperOrigin-RevId: 815183251