[ObjC] Fix nullability annotations on some `GPB*Dictionary` types.
As announced: https://protobuf.dev/news/2025-09-19/#objc

Correcting the some of the annotations as part of ObjC Major Version
bump coming in 2026Q1.

This is most likely to trip up Swift code calling it, but it could also
impact ObjC sources depending on the annotation enforcement the files
are compiled with.

Fixes #22403

PiperOrigin-RevId: 815778859