Update Rust runtime version check to reflect that v34 will be the first stable release
Note that this change would break the v33 release if it landed on the 33.x
branch. This is because the logic now treats v33 gencode as incompatible with
the runtime. However, this is fine since the 33.x branch cut has already been
done and we are not going to cherry-pick this change.

PiperOrigin-RevId: 816785840