[ObjC] Fix some tests to not violate GPBCodedOutputStream invariants.
If the class is writing to an NSOutputStream, the buffer must always
be used to batch things up. The tests were accidentally making a zero
length buffer which isn't valid.

Fix the tests to not do this.

Make the core code assert in debug to ensure the invariant can't be
broken.

PiperOrigin-RevId: 817153199