descriptor: Expose DescriptorBuilder::IsEnumNamespaceMessage.
This functionality is useful for static analysis tooling and is difficult to
re-implement because it needs access to FieldDescriptor::kLastReservedNumber.

The implementation does have to make some compromises now that it no longer has
access to private members of Descriptor and EnumDescriptor, but fortunately the
needed data is cheaply available in proto form.

PiperOrigin-RevId: 817758807