Take the arena pointer as an argument to offset-enabled constructors of fields.
In some offset-enabled field constructors, the arena pointer is needed (e.g. copy constructors). Rather than derive the arena from the arena offset, we can pass the arena pointer from above directly to save the cost of `GetArena()`. We already have this arena pointer in all callers of offset-enabled constructors, so this has no cost for callers that don't use the arena.

PiperOrigin-RevId: 819398420