Fix Java IterableByteBufferInputStream to return -1 on empty
This gives the InputStream slightly less work to do when it knows there
are no buffers that can be checked for more data.

Also loosened a check in CodedInputStream to permit read() to return
zero, as this is documented as a valid value - though in this case -1 is
the correct value as there is nothing more to read.

Fixes #23957