Avoid calling `GetArena()` in `~RepeatedField` when the element is arena-skippable.
The destructor of `RepeatedField` will never be called if allocated on an arena with a destructor-skippable element type. In this case, we know the arena pointer is null, and we avoid a call to `RepeatedField::GetArena()`.

`InternalDeallocate` does not actually use the arena pointer when `in_destructor` is set. However, as the current implementation stands, checking whether the arena pointer is `nullptr` from a field with an arena offset requires reading the arena pointer from the `InternalMetadata`, since it is possible to have a non-sentinel offset and a `nullptr` arena. I will later be experimenting with an optimization that will reduce this cost when we only care about whether an arena is present.

PiperOrigin-RevId: 820538848